+++ exchangelib/folders/base.py
@@ -6,7 +6,7 @@ from operator import attrgetter
 from .collections import FolderCollection, SyncCompleted
 from .queryset import SingleFolderQuerySet, SHALLOW as SHALLOW_FOLDERS, DEEP as DEEP_FOLDERS
 from ..errors import ErrorAccessDenied, ErrorFolderNotFound, ErrorCannotEmptyFolder, ErrorCannotDeleteObject, \
-    ErrorDeleteDistinguishedFolder, ErrorInvalidSubscription, ErrorNoPublicFolderReplicaAvailable, ErrorItemNotFound
+    ErrorDeleteDistinguishedFolder, ErrorNoPublicFolderReplicaAvailable, ErrorItemNotFound
 from ..fields import IntegerField, CharField, FieldPath, EffectiveRightsField, PermissionSetField, EWSElementField, \
     Field, IdElementField, InvalidField
 from ..items import CalendarItem, RegisterMixIn, ITEM_CLASSES, DELETE_TYPE_CHOICES, HARD_DELETE, \
@@ -681,8 +681,6 @@ class BaseFolder(RegisterMixIn, SearchableMixIn, metaclass=EWSMeta):
             if max_notifications_returned and i >= max_notifications_returned:
                 svc.stop_streaming()
                 break
-        if svc.error_subscription_ids:
-            raise ErrorInvalidSubscription('Invalid subscription IDs: %s' % svc.error_subscription_ids)
 
     def __floordiv__(self, other):
         """Support the some_folder // 'child_folder' // 'child_of_child_folder' navigation syntax.